Replace 'Flatcar Linux' with 'Flatcar Container Linux'
As this is official name we should be using where possible.

Name of the project stays as it is for now on to avoid confusion and
creating problems. Also FLUO sounds nicer than FCLUO.

This commit also changes FilterContainerLinuxNodes() function, which
selects Flatcar nodes from list of Kubernetes nodes based on osImage
field from nodeInfo, which is taken from PRETTY_NAME field in /etc/os-release
file on each node.

This field has been changed from 'Flatcar Linux' to 'Flatcar Container
Linux' in 2247.7.0 Flatcar version, which was released almost year ago,
so since then the function was broken.

As this change has been released upstream long time ago, it should be okay
to drop support for the old variant.

Closes #25
Closes #28

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>