pkg/k8sutil: fail if we can't read update configuration override file
Mainly to avoid calling Close() manually but also to ensure correctness,
this commit changes getUpdateMap() method to return error, if
configuration file for update mechanism exists, but it's not readable.

If the file does not exist, we continue as usual.

This commit also adds an internal unit test to splitNewlineEnv()
function to make sure that when given empty input, it retains the values
in the map.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>