pkg/operator: shutdown leader election when control channel is closed
To avoid leaking goroutine and to exit cleanly when user requests it,
instead of panicking. If user did not request shutdown and we lose
leadership, runtime error is returned, so update-operator binary should
exit with non-zero exit code, which preserves current behavior. After
that, process will be restarted and will be waiting for leader election,
as before.

Closes #80

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>