pkg/operator: pass around context to all operations requiring it
Right now we create a top-level context which never times out and we
simply pass it around to avoid changing existing behavior.

No timeout shouldn't be a problem, as right now the controller loops
runs quickly and once control channel is closed, it won't be executed
next time, so controller will exit gracefully.

However, adding global context will allows us later on to for example
add some graceful shutdown period and it resolves use of context.TODO(),
which should not really be used in maintained code.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>