pkg/agent: check Run context when waiting for ok to reboot
This way, when context is cancelled, we exit the loop gracefully, rather
than hanging infinitely.

No error returned in this case and is arbitrary, it could as well be an
error. However, we actually really wait for both reboot need and reboot
permission here, so exiting without error likely make sense here.

Signed-off-by: Mateusz Gozdek <mgozdek@microsoft.com>