pkg/k8sutil: use fake client instead of gomock for testing
This makes this tests more aligned with rest of the codebase and
eliminates the need of using gomock completely, so we can drop code
generation from the codebase which adds unnecessary complexity.

Also using fake client is more type safe and natural as oppose to
gomock.

This commit also adds handling for more cases for function updating the
node by following property based testing.

Signed-off-by: Mateusz Gozdek <mgozdek@microsoft.com>