Drop dependency on locksmith
As we need to update it and new version requires 'replace' statements
for etcd dependency, which I would like to avoid.

Given that we only depend on the periodic code, this is now being copied
into the repository and refactored a bit to fix found linter issues.

See #171 for more details.

Signed-off-by: Mateusz Gozdek <mgozdek@microsoft.com>